Add files via upload
Demo of how to generate dummy data using faker and pandas libraries.
This dummy dataset has 4 columns with headers and this is going to generate 100 rows in each file and there will be 15 files generated in the output location mentioned within the code. You can change the no of files been generated by increase/decrease the no of times the for loop runs.