Add files via upload
In this demo we get stream of data files, which is generated by dummy data generator(Notebook_9). So please run dummy data generator code first and then spark submit this data-streaming demo.py file to be able to read that data and then process it.
That stream of data is grouped by device_id to aggregate the count of occurrences for a window of 5 sec.
And that aggregated data along with the windowed timestamp is outputted on the console.